requirements.txt
requests
Crypto
paho-mqtt
